requirements.txt
packaging >= 23.0
PyGithub >= 1.58
